Fix possible lossy conversion

+++ NT_el3vate/src/rw_primitive.cpp
@@ -261,7 +261,7 @@ unsigned __int64 GetEPROCESSPhysicalBase(const char* processName ,int pid ,HANDL
 				offset_into_mapped_area = 0;
 				int correct_MEMORY_MAPED_SIZE = MEMORY_MAPED_SIZE;
 				if (page + MEMORY_MAPED_SIZE > end) {
-					correct_MEMORY_MAPED_SIZE = MEMORY_MAPED_SIZE - (page + MEMORY_MAPED_SIZE - end);
+					correct_MEMORY_MAPED_SIZE = MEMORY_MAPED_SIZE - (unsigned __int64)(page + MEMORY_MAPED_SIZE - end);
 				}
 				if (MapPhysicalMemory((HANDLE) * (PDWORD64)hPhysicalMemory, page, correct_MEMORY_MAPED_SIZE, buf) == FALSE) {
 					fprintf(stderr, "[!] MapPhysicalMemory failed\n");
